requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
argon2-cffi=21.1.0=py36h8f6f2f9_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=5.0.1=pyhd8ed1ab_0
blosc=1.21.0=h8c45485_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.14.6=py36hc120d54_0
charls=2.1.0=he6710b0_2
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
clip=1.0=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.7.99=0
cuda-gdb=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2021.3.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=py36_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.4.1=pypi_0
fontconfig=2.13.1=hef1e5e3_1
freetype=2.12.1=h4a9f257_0
ftfy=5.8=py_0
gds-tools=1.4.0.31=0
giflib=5.1.4=h14c3975_1
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
huggingface-hub=0.4.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imagecodecs=2020.5.30=py36hfa7d478_2
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.1.0=h9e868ea_3769
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.7.2=pyhd8ed1ab_0
jedi=0.17.0=py36_0
jinja2=2.11.3=pyhd8ed1ab_2
joblib=1.1.1=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=3.0.2=py36_0
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.8.1=py36h06a4308_0
jupyterlab_widgets=1.1.1=pyhd8ed1ab_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libaec=1.0.4=he6710b0_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.0.1=h8e7db2f_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
libzopfli=1.0.3=he6710b0_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py36h8f6f2f9_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mistune=0.8.4=py36h8f6f2f9_1004
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
nbconvert=5.6.1=pyhd8ed1ab_2
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
networkx=2.5=py_0
nltk=3.4.5=py36_0
notebook=6.3.0=py36h5fab9bb_0
nsight-compute=2022.3.0.22=0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
prometheus_client=0.15.0=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.17.3=py36h8f6f2f9_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.6=2_cp36m
pytorch=1.10.2=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-mutex=1.0=cuda
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=py36h27cfd23_1
pyzmq=22.2.1=py36h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.2=h5eee18b_0
regex=2021.8.3=py36h7f8727e_0
requests=2.27.1=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.17.2=py36hdf5156a_0
scipy=1.2.1=py36h7c811a0_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.40.0=h5082296_0
terminado=0.12.1=py36h5fab9bb_0
testpath=0.6.0=pyhd8ed1ab_0
tifffile=2021.3.17=pyhd3eb1b0_1
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
toolz=0.11.2=pyhd3eb1b0_0
torch-summary=1.4.5=pypi_0
torchvision=0.11.3=py36_cu113
tornado=6.1=py36h27cfd23_0
tqdm=4.63.0=pyhd3eb1b0_0
traitlets=4.3.3=py36h06a4308_0
transformers=4.18.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.13=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pyha770c72_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
